refactor(example): migrate to golib gormdao and breaking-change APIs
