+++ cmd/generate/example/pkg/code/code.go
@@ -3,8 +3,7 @@ package code
 import (
 	"fmt"
 
-	"github.com/morehao/golib/biz/gconstant"
-	"github.com/morehao/golib/biz/genericdao"
+	"github.com/morehao/golib/gconstant"
 	"github.com/morehao/golib/gerror"
 )
 
@@ -31,7 +30,7 @@ func GetError(code int) *gerror.Error {
 func init() {
 	// 业务错误码规范: 从 1002XX 开始
 	// 模块划分: 1002XX(租户) 1003XX(公司) 1004XX(部门) 1005XX(用户) 1006XX(菜单) 1007XX(角色)
-	registerError(genericdao.DBErrorMsgMap)
+	registerError(gconstant.DBErrorMsgMap)
 	registerError(gconstant.SystemErrorMsgMap)
 	registerError(gconstant.AuthErrorMsgMap)
 	registerError(userErrorMsgMap)
