+++ cmd/generate/example/pkg/dbclient/gorm.go
@@ -5,8 +5,9 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/morehao/golib/biz/gormplugin"
+	_ "github.com/morehao/golib/dbaccess/dbgorm/driver/mysql"
 	"github.com/morehao/golib/dbaccess/dbgorm"
+	"github.com/morehao/golib/dbaccess/gormplugin"
 	"github.com/morehao/golib/glog"
 	"gorm.io/gorm"
 )
@@ -21,7 +22,7 @@ const (
 	dbNameIam  = "ark_iam"
 )
 
-func InitMultiDB(configs []dbgorm.GormConfig, logConfig *glog.LogConfig) error {
+func InitMultiDB(configs []dbgorm.Config, logConfig *glog.LogConfig) error {
 	if len(configs) == 0 {
 		return fmt.Errorf("mysql config is empty")
 	}
